+++ src/main.rs
@@ -2,20 +2,10 @@
 use std::env;
 use std::sync::{Arc, Mutex};
 
-use csv::{self};
-use std::error::Error;
-use std::str::FromStr;
-
 use nalgebra::DMatrix;
-use optimizers::rmsprop::RMSProp;
 
-use crate::{
-    core::{layer::Layer, model::Model},
-    functions::{
-        activations::{relu, relu_derivative, softmax, softmax_derivative},
-        losses::{categorical_crossentropy, categorical_crossentropy_derivative},
-    },
-};
+use crate::core::model::Model;
+use crate::model_handler::get_trained_model;
 
 use actix_cors::Cors;
 use actix_web::{web, App, HttpResponse, HttpServer, Responder};
@@ -23,68 +13,8 @@ use serde::{Deserialize, Serialize};
 
 mod core;
 mod functions;
+mod model_handler;
 mod optimizers;
-
-fn read_mnist(file_path: &str) -> Result<(Vec<DMatrix<f32>>, Vec<DMatrix<f32>>), Box<dyn Error>> {
-    let mut rdr = csv::Reader::from_path(file_path)?;
-
-    let mut x: Vec<DMatrix<f32>> = Vec::new();
-    let mut y: Vec<DMatrix<f32>> = Vec::new();
-
-    for result in rdr.records() {
-        let record = result?;
-
-        let class = record[0].parse::<usize>().unwrap();
-
-        let float_record: Vec<f32> = record
-            .iter()
-            .map(|x| x.parse::<f32>().unwrap() / 255.0)
-            .collect();
-
-        x.push(DMatrix::from_vec(
-            float_record.len() - 1,
-            1,
-            float_record[1..].to_vec(),
-        ));
-
-        let mut label = DMatrix::zeros(1, 10);
-
-        label[class] = 1.0;
-
-        y.push(label.transpose());
-    }
-
-    Ok((x, y))
-}
-
-fn read_doodles(file_path: &str) -> Result<(Vec<DMatrix<f32>>, Vec<DMatrix<f32>>), Box<dyn Error>> {
-    let mut rdr = csv::Reader::from_path(file_path)?;
-
-    let mut x: Vec<DMatrix<f32>> = Vec::new();
-    let mut y: Vec<DMatrix<f32>> = Vec::new();
-
-    for result in rdr.records() {
-        let record = result?;
-
-        let class = record[1].parse::<usize>().unwrap();
-
-        let numbers: Vec<f32> = record[2]
-            .trim_matches(|p| p == '[' || p == ']')
-            .split(',')
-            .map(|s| f32::from_str(s.trim()).unwrap() / 255.0)
-            .collect();
-
-        x.push(DMatrix::from_vec(numbers.len(), 1, numbers));
-
-        let mut label = DMatrix::zeros(1, 10);
-
-        label[class] = 1.0;
-
-        y.push(label.transpose());
-    }
-
-    Ok((x, y))
-}
 
 #[derive(Deserialize)]
 struct MatrixData {
@@ -111,15 +41,13 @@ async fn upload_matrix(
         }
     }
 
-    let dmatrix = DMatrix::from_row_slice(matrix_elements.len(), 1, &matrix_elements);
-    // Use `dmatrix` as needed
+    let input_data = DMatrix::from_row_slice(matrix_elements.len(), 1, &matrix_elements);
     let mut locked_model = shared.lock().unwrap();
 
-    let prediction = locked_model.evaluate(&dmatrix);
+    let prediction = locked_model.evaluate(&input_data);
 
     drop(locked_model);
 
-    // println!("Model predicted: {}", prediction);
     let prediction_vec = prediction.data.as_vec();
     let mut index_of_max: Option<usize> = prediction_vec
         .iter()
@@ -141,66 +69,13 @@ async fn upload_matrix(
 async fn main() -> std::io::Result<()> {
     env::set_var("RUST_BACKTRACE", "1");
 
-    let model = match (
-        read_doodles("./doodles/train-quick-draw.csv"),
-        read_doodles("./doodles/test-quick-draw.csv"),
-    ) {
-        (Ok((x_train, y_train)), Ok((x_test, y_test))) => {
-            println!(
-                "Loaded data x_train = {} y_train = {} x_test = {} y_test = {})",
-                x_train.len(),
-                y_train.len(),
-                x_test.len(),
-                y_test.len()
-            );
-            println!("x[0] = {}", x_train[0]);
+    println!("Training the network...");
 
-            let hidden_layer1 = Layer::new(relu, relu_derivative, x_train[0].len(), 512);
-
-            let hidden_layer2 = Layer::new(relu, relu_derivative, 512, 512);
-
-            let output_layer = Layer::new(softmax, softmax_derivative, 512, y_train[0].len());
-
-            let mut model = Model::new(
-                vec![hidden_layer1, hidden_layer2, output_layer],
-                categorical_crossentropy,
-                categorical_crossentropy_derivative,
-            );
-
-            let mut rmsprop = RMSProp::new(0.9);
-
-            let metrics = vec![
-                "accuracy".to_string(),
-                "recall".to_string(),
-                "f1-score".to_string(),
-                "precision".to_string(),
-            ];
-
-            model.fit(
-                128,
-                15,
-                0.001,
-                metrics.clone(),
-                &mut rmsprop,
-                x_train,
-                y_train,
-            );
-
-            println!("\nTest results:\n");
-
-            model.test(metrics, &x_test, &y_test);
-
-            Some(model)
-        }
-        _ => {
-            println!("Error reading CSV file:");
-            None
-        }
-    };
+    let model = get_trained_model();
 
     let shared_data = Arc::new(Mutex::new(model.unwrap()));
 
-    println!("Started API");
+    println!("\n===== Started API =====");
 
     HttpServer::new(move || {
         let cors = Cors::default()
